fix inferer and lint
